+++ frontend/src/App.tsx
@@ -5,11 +5,12 @@ import Footer from "./components/Footer"
 import Home from "./pages/home/Home"
 import About from "./pages/About"
 import Partners from "./pages/partners/PartnersPage"
-import Contact from "./pages/contact/Contact"
+import Contact from "./pages/contact/ContactPage"
 import Login from "./pages/login/Login"
 import Register from "./pages/login/Register"
 import Catalog from "./pages/products/CatalogPage"
 import Cart from "./pages/cart/Cart"
+import Checkout from "./pages/Checkout"
 import NotFound from "./pages/NotFound"
 import "./locales/i18n"
 import "./App.css"
@@ -40,11 +41,12 @@ function App() {
         <Route path="/:lang/register" element={<Register />} />
         <Route path="/:lang/products" element={<Catalog />} />
         <Route path="/:lang/cart" element={<Cart />} />
+        <Route path="/:lang/checkout" element={<Checkout />} />
 
         {/* 404 route */}
         <Route path="*" element={<NotFound />} />
       </Routes>
-      {/* <Footer /> */}
+      <Footer />
     </Router>
   )
 }