+++ frontend/src/pages/cart/Cart.tsx
@@ -1,7 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { useState, useEffect } from "react";
 import { Container, ListGroup, Row, Col } from "react-bootstrap";
-import { useNavigate } from "react-router-dom";
 import CartItem from "./Item";
 import ToastNotification from "@/components/Toast";
 import CartOverview from "./Overview";
@@ -17,7 +16,6 @@ import {
 
 const Cart = () => {
   const { t } = useTranslation();
-  const navigate = useNavigate();
 
   // State for cart items and toasts
   const [cart, setCart] = useState<any[]>([]);
@@ -31,19 +29,14 @@ const Cart = () => {
     loadCart(setCart);
   }, []);
 
-  // Calculate Total Price and Item Count
+  // Recalculate totals whenever the cart updates
   const totalItems = calculateTotalItems(cart);
   const subtotal = calculateSubtotal(cart);
   const totalPrice = subtotal + (subtotal > 0 ? SHIPPING_COST : 0);
 
-  // Handle Checkout
-  const handleCheckout = () => {
-    navigate("/checkout");
-  };
-
   return (
     <Container>
-      <h2 className="cart-title text-center mb-4">{t("cart.title")}</h2>
+      <h2 className="cart-title text-center mb-5">{t("cart.title")}</h2>
 
       {/* Toast Notifications */}
       <ToastNotification
@@ -63,15 +56,15 @@ const Cart = () => {
                 <ListGroup.Item key={item.id} className="mb-3 p-3">
                   <CartItem
                     item={item}
-                    onIncrease={() =>
-                      handleIncreaseQuantity(item.id, cart, setCart)
-                    }
-                    onDecrease={() =>
-                      handleDecreaseQuantity(item.id, cart, setCart)
-                    }
-                    onRemove={() =>
-                      handleRemoveItem(item.id, cart, setCart, setToasts, t)
-                    }
+                    onIncrease={() => {
+                      handleIncreaseQuantity(item.id, cart, setCart);
+                    }}
+                    onDecrease={() => {
+                      handleDecreaseQuantity(item.id, cart, setCart);
+                    }}
+                    onRemove={() => {
+                      handleRemoveItem(item.id, cart, setCart, setToasts, t);
+                    }}
                   />
                 </ListGroup.Item>
               ))}
@@ -79,13 +72,12 @@ const Cart = () => {
           </Col>
 
           {/* Overview Section */}
-          <Col md={4}>
+          <Col md={4} className="overview-top-right">
             <CartOverview
               totalItems={totalItems}
               subtotal={subtotal}
               totalPrice={totalPrice}
               shippingCost={SHIPPING_COST}
-              onCheckout={handleCheckout}
             />
           </Col>
         </Row>