+++ frontend/src/pages/cart/Overview.tsx
@@ -1,12 +1,13 @@
 import { useTranslation } from "react-i18next";
 import { ListGroup, Card, Button } from "react-bootstrap";
+import { useParams, useNavigate } from "react-router-dom";
 
+// Props interface for CartOverview
 interface CartOverviewProps {
   totalItems: number;
   subtotal: number;
   totalPrice: number;
   shippingCost: number;
-  onCheckout: () => void;
 }
 
 const CartOverview: React.FC<CartOverviewProps> = ({
@@ -14,9 +15,10 @@ const CartOverview: React.FC<CartOverviewProps> = ({
   subtotal,
   totalPrice,
   shippingCost,
-  onCheckout,
 }) => {
   const { t } = useTranslation();
+  const { lang } = useParams();
+  const navigate = useNavigate();
 
   return (
     <Card className="p-3 mb-3">
@@ -24,26 +26,22 @@ const CartOverview: React.FC<CartOverviewProps> = ({
       <ListGroup variant="flush">
         <ListGroup.Item className="d-flex justify-content-between">
           <span>{t("cart.itemCount")}</span>
-          <strong>{totalItems}</strong>
+          {totalItems}
         </ListGroup.Item>
         <ListGroup.Item className="d-flex justify-content-between">
           <span>{t("cart.subtotal")}</span>
-          <strong>
-            {t("currency")}
-            {subtotal.toFixed(2)}
-          </strong>
+          {t("currency")}
+          {subtotal.toFixed(2)}
         </ListGroup.Item>
         <ListGroup.Item className="d-flex justify-content-between">
           <span>{t("cart.shipping")}</span>
-          <strong>
-            {subtotal > 0
-              ? `${t("currency")}${shippingCost.toFixed(2)}`
-              : t("cart.free")}
-          </strong>
+          {subtotal > 0
+            ? `${t("currency")}${shippingCost.toFixed(2)}`
+            : t("cart.free")}
         </ListGroup.Item>
         <ListGroup.Item className="d-flex justify-content-between">
           <span>{t("cart.total")}</span>
-          <strong>
+          <strong style={{ textDecoration: "underline" }}>
             {t("currency")}
             {totalPrice.toFixed(2)}
           </strong>
@@ -54,7 +52,7 @@ const CartOverview: React.FC<CartOverviewProps> = ({
         size="lg"
         className="mt-3 w-100"
         disabled={totalItems === 0}
-        onClick={onCheckout}
+        onClick={() => navigate(`/${lang}/checkout`)}
       >
         {t("cart.checkout")}
       </Button>