+++ frontend/src/pages/home/Home.tsx
@@ -1,8 +1,7 @@
-import Carousel from "./Carousel"
-import Section from "./Section"
-import Partners from "../partners/Partners"
-import ProductCatalog from "../products/ProductCatalog"
-import Footer from "@/components/Footer"
+import Carousel from "./Carousel";
+import Section from "./Section";
+import Partners from "../partners/Partners";
+import ProductCatalog from "../products/ProductCatalog";
 
 const Home = () => {
   return (
@@ -11,10 +10,8 @@ const Home = () => {
       <Section />
       <ProductCatalog showDetails={false} />
       <Partners />
-      <br />
-      <Footer />
     </>
-  )
-}
+  );
+};
 
-export default Home
+export default Home;