+++ frontend/src/pages/products/CatalogPage.tsx
@@ -1,22 +1,18 @@
-import { useEffect } from "react"
-import Footer from "@/components/Footer"
-import ProductCatalog from "./ProductCatalog"
+import { useEffect } from "react";
+import ProductCatalog from "./ProductCatalog";
 
 const CatalogPage = () => {
   // Scroll to the top of the page when the component mounts
   useEffect(() => {
-    window.scrollTo(0, 0)
-  }, [])
+    window.scrollTo(0, 0);
+  }, []);
 
   // Render the Catalog component
   return (
     <div className="d-flex flex-column" style={{ minHeight: "100vh" }}>
-      <main className="mb-5" style={{ flexGrow: 1 }}>
-        <ProductCatalog showDetails={true} />
-      </main>
-      <Footer />
+      <ProductCatalog showDetails={true} />
     </div>
-  )
-}
+  );
+};
 
-export default CatalogPage
+export default CatalogPage;